pyproject.toml
[project]
name = "legal-check"
version = "0.1.0"
description = ""
authors = [{ name = "dobleuber", email = "wbertc@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = ["langchain (>=0.3.14,<0.4.0)", "langchain-community (>=0.3.14,<0.4.0)", "pypdf (>=5.1.0,<6.0.0)", "pydantic (>=2.10.5,<3.0.0)", "langchain-openai (>=0.3.1,<0.4.0)", "beautifulsoup4 (>=4.12.3,<5.0.0)", "python-dotenv (>=1.0.1,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "legal_check:main"

